+++ examples/vrpc-react-todos-2/test/test.sh
@@ -50,6 +50,6 @@ else
   printf "${GREEN}Tests Passed (${CONTAINER})${NC}\n"
 fi
 # call the cleanup fuction
-#cleanup
+cleanup
 # exit the script with the same code as the test service code
 exit $TEST_EXIT_CODE